perf(rust): land custom Deserialize + float-field hints + snake_case fix
Three coordinated correctness/perf changes that together turn the
parse time from ~268ms (baseline) to ~40ms (~470k entries/sec).

1. Custom Deserialize for MasterfileEntry. Dispatches by peeking at
   the inner `data` object's non-templateId key — O(1) match instead
   of O(n_variants) untagged scan. Stub entries (3 of them) dispatch
   by templateId value match.

2. Extends scanFloatFields to also catch decimals inside arrays-of-
   numbers (e.g., `buddyPortraitOffset: [-250.0, 0.0, 0.0]`). The
   inference layer propagates the float hint through array/tuple
   element types via applyFloatToNumbers, stopping at object
   boundaries (each field has its own hint lookup).

3. Fixes snakeCase to handle the uppercase-run-before-PascalWord
   boundary. JSON has fields like `maxStoneACount` and `parseHTML
   Document`; the old converter produced `max_stone_acount` (which
   round-tripped through serde rename_all=camelCase as
   `maxStoneAcount` — a different JSON field). Now correctly produces
   `max_stone_a_count` → `maxStoneACount`.

Without #2 and #3, custom Deserialize exposes the data-vs-type
mismatches that the auto-derive untagged was silently masking by
falling back to permissive stub variants. With them in, dispatch
hits the right variant on the first try and types match the actual
JSON.

Verified: cargo check clean, examples/parse parses all 18,239
entries cleanly. Three runs: 40ms, 38ms, 40ms.

+++ packages/rust/src/move_settings.rs
@@ -15,7 +15,7 @@ pub struct MoveSettings {
     pub heal_scalar: Option<f64>,
     pub is_locked: Option<bool>,
     pub movement_id: String,
-    pub ob_move_settings_number18: Option<[u64; 4]>,
+    pub ob_move_settings_number18: Option<[f64; 4]>,
     pub pokemon_type: String,
     pub power: Option<f64>,
     pub stamina_loss_scalar: Option<f64>,
